+++ python/ferrmion/hamiltonians/__init__.py
@@ -1,0 +1,18 @@
+"""Initialisation for Hamiltonians."""
+
+from .molecular import molecular_hamiltonian, molecular_hamiltonian_template
+from .utils import (
+    fill_template,
+    symplectic_product_map,
+    to_qubit_hamiltonian,
+    to_symplectic_hamiltonian,
+)
+
+__all__ = [
+    "molecular_hamiltonian_template",
+    "molecular_hamiltonian",
+    "fill_template",
+    "to_qubit_hamiltonian",
+    "to_symplectic_hamiltonian",
+    "symplectic_product_map",
+]