+++ python/tests/test_ternary_tree.py
@@ -6,27 +6,12 @@ from ferrmion.utils import symplectic_hash, symplectic_unhash
 from openfermion import QubitOperator, get_sparse_operator
 from openfermion.ops import InteractionOperator
 from openfermion.transforms import jordan_wigner
+from ferrmion.hamiltonians.molecular import molecular_hamiltonian
 
 
 @pytest.fixture
-def one_e_ints():
-    return np.random.random((6, 6))
-
-
-@pytest.fixture
-def two_e_ints():
-    return np.random.random((6, 6, 6, 6))
-
-
-@pytest.fixture
-def fermion_modes():
-    return {i for i in range(6)}
-
-
-@pytest.fixture
-def six_mode_tree(one_e_ints, two_e_ints):
-    tt = TernaryTree(one_e_ints, two_e_ints, root_node=TTNode())
-    return tt
+def six_mode_tree():
+    return TernaryTree(n_modes=6, root_node=TTNode())
 
 
 def test_ttnode():
@@ -296,10 +281,7 @@ def test_bravyi_kitaev(six_mode_tree):
 
 
 def tests_bonsai_paper_tree():
-    tt = TernaryTree(
-        np.zeros((11, 11)),
-        np.zeros((11, 11, 11, 11)),
-    )
+    tt = TernaryTree(n_modes=11)
     tt = tt.add_node("x")
     tt = tt.add_node("y")
     tt = tt.add_node("z")
@@ -455,19 +437,25 @@ def tests_bonsai_paper_tree():
     for line in tt._build_symplectic_matrix()[1]:
         assert np.all(line == symplectic_unhash(symplectic_hash(line), len(line)))
 
-
-def test_eigenvalues_with_openfermion(six_mode_tree):
+def test_eigenvalues_with_openfermion(water_eigenvalues, water_integrals):
     # qham_zeros = InteractionOperator(0, tt.one_e_coeffs, np.zeros(tt.two_e_coeffs.shape))
     # ofop_zeros = jordan_wigner(qham_zeros)
+    one_e_ints, two_e_ints = water_integrals
     qham = InteractionOperator(
-        0, six_mode_tree.one_e_coeffs, six_mode_tree.two_e_coeffs
+        0, one_e_ints, 0.5*two_e_ints
     )
     # print(qham)
     ofop = jordan_wigner(qham)
     # print(f"diff {ofop-ofop_zeros}")
     diag, _ = sp.sparse.linalg.eigsh(get_sparse_operator(ofop), k=6, which="SA")
 
-    qham2 = six_mode_tree.JW().to_qubit_hamiltonian()
+    assert np.allclose(sorted(diag), sorted(water_eigenvalues))
+
+
+def test_eigenvalues_across_encodings(water_eigenvalues, water_tt, water_integrals):
+    one_e_ints, two_e_ints = water_integrals
+
+    qham2 = molecular_hamiltonian(water_tt.JKMN(), one_e_ints, 0.5*two_e_ints, 0)
     ofop2 = QubitOperator()
     for k, v in qham2.items():
         string = " ".join(
@@ -479,38 +467,10 @@ def test_eigenvalues_with_openfermion(six_mode_tree):
         ofop2 += QubitOperator(term=string, coefficient=v)
     diag2, _ = sp.sparse.linalg.eigsh(get_sparse_operator(ofop2), k=6, which="SA")
 
-    assert np.allclose(diag, diag2)
+    assert np.allclose(sorted(water_eigenvalues), sorted(diag2))
 
 
-def test_eigenvalues_across_encodings(six_mode_tree):
-    qham = six_mode_tree.JW().to_qubit_hamiltonian()
-    ofop = QubitOperator()
-    for k, v in qham.items():
-        string = " ".join(
-            [
-                f"{char.upper()}{pos}" if char != "I" else ""
-                for pos, char in enumerate(k)
-            ]
-        )
-        ofop += QubitOperator(term=string, coefficient=v)
-    diag, _ = sp.sparse.linalg.eigsh(get_sparse_operator(ofop), k=6, which="SA")
-
-    qham2 = six_mode_tree.JKMN().to_qubit_hamiltonian()
-    ofop2 = QubitOperator()
-    for k, v in qham2.items():
-        string = " ".join(
-            [
-                f"{char.upper()}{pos}" if char != "I" else ""
-                for pos, char in enumerate(k)
-            ]
-        )
-        ofop2 += QubitOperator(term=string, coefficient=v)
-    diag2, _ = sp.sparse.linalg.eigsh(get_sparse_operator(ofop2), k=6, which="SA")
-
-    assert np.allclose(sorted(diag), sorted(diag2))
-
-
-def test_default_mode_op_map(six_mode_tree):
-    assert six_mode_tree.default_mode_op_map == {
-        i: i for i in range(six_mode_tree.n_qubits)
+def test_default_mode_op_map(water_tt):
+    assert water_tt.default_mode_op_map == {
+        i: i for i in range(water_tt.n_qubits)
     }